Fix some signedness warnings.

+++ mkfs/main.c
@@ -121,9 +121,9 @@ static int setup_spc_bits(int sector_bits, int user_defined, off_t volume_size)
 		return user_defined;
 	}
 
-	if (volume_size < 256ull * 1024 * 1024)
+	if (volume_size < 256LL * 1024 * 1024)
 		return MAX(0, 12 - sector_bits);	/* 4 KB */
-	if (volume_size < 32ull * 1024 * 1024 * 1024)
+	if (volume_size < 32LL * 1024 * 1024 * 1024)
 		return MAX(0, 15 - sector_bits);	/* 32 KB */
 
 	for (i = 17; ; i++)						/* 128 KB or more */
@@ -178,7 +178,7 @@ static int setup(struct exfat_dev* dev, int sector_bits, int spc_bits,
 
 static int logarithm2(int n)
 {
-	int i;
+	size_t i;
 
 	for (i = 0; i < sizeof(int) * CHAR_BIT - 1; i++)
 		if ((1 << i) == n)
